:wrench: chore: move notif utils out of __init__.py to prevent circular imports
