+++ src/sentry/notifications/utils/__init__.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import logging
-import time
 from collections import defaultdict
 from collections.abc import Iterable, Mapping, MutableMapping, Sequence
 from dataclasses import dataclass
@@ -10,12 +9,10 @@ from typing import TYPE_CHECKING, Any, Optional, TypedDict, Union, cast
 from urllib.parse import parse_qs, urlparse
 
 from django.db.models import Count
-from django.utils.http import urlencode
 from django.utils.safestring import mark_safe
 from django.utils.translation import gettext_lazy as _
 
 from sentry.eventstore.models import Event, GroupEvent
-from sentry.incidents.models.alert_rule import AlertRuleTriggerAction
 from sentry.integrations.base import IntegrationFeatures, IntegrationProvider
 from sentry.integrations.manager import default_manager as integrations
 from sentry.integrations.services.integration import integration_service
@@ -42,6 +39,8 @@ from sentry.utils.committers import get_serialized_event_file_committers
 from sentry.utils.performance_issues.base import get_url_from_span
 from sentry.utils.performance_issues.performance_problem import PerformanceProblem
 from sentry.web.helpers import render_to_string
+
+from .types import NotificationRuleDetails
 
 if TYPE_CHECKING:
     from sentry.notifications.notifications.activity.base import ActivityNotification
@@ -116,90 +115,6 @@ def get_environment_for_deploy(deploy: Deploy | None) -> str:
         if environment and environment.name:
             return str(environment.name)
     return "Default Environment"
-
-
-def get_email_link_extra_params(
-    referrer: str = "alert_email",
-    environment: str | None = None,
-    rule_details: Sequence[NotificationRuleDetails] | None = None,
-    alert_timestamp: int | None = None,
-    notification_uuid: str | None = None,
-    **kwargs: Any,
-) -> dict[int, str]:
-    alert_timestamp_str = (
-        str(round(time.time() * 1000)) if not alert_timestamp else str(alert_timestamp)
-    )
-    return {
-        rule_detail.id: "?"
-        + str(
-            urlencode(
-                {
-                    "referrer": referrer,
-                    "alert_type": str(AlertRuleTriggerAction.Type.EMAIL.name).lower(),
-                    "alert_timestamp": alert_timestamp_str,
-                    "alert_rule_id": rule_detail.id,
-                    **dict(
-                        []
-                        if notification_uuid is None
-                        else [("notification_uuid", str(notification_uuid))]
-                    ),
-                    **dict([] if environment is None else [("environment", environment)]),
-                    **kwargs,
-                }
-            )
-        )
-        for rule_detail in (rule_details or [])
-    }
-
-
-def get_group_settings_link(
-    group: Group,
-    environment: str | None,
-    rule_details: Sequence[NotificationRuleDetails] | None = None,
-    alert_timestamp: int | None = None,
-    referrer: str = "alert_email",
-    notification_uuid: str | None = None,
-    **kwargs: Any,
-) -> str:
-    alert_rule_id = rule_details[0].id if rule_details and rule_details[0].id else None
-    extra_params = ""
-    if alert_rule_id:
-        extra_params = get_email_link_extra_params(
-            referrer,
-            environment,
-            rule_details,
-            alert_timestamp,
-            notification_uuid=notification_uuid,
-            **kwargs,
-        )[alert_rule_id]
-    elif not alert_rule_id and notification_uuid:
-        extra_params = "?" + str(urlencode({"notification_uuid": notification_uuid}))
-    return str(group.get_absolute_url() + extra_params)
-
-
-def get_integration_link(
-    organization: Organization, integration_slug: str, notification_uuid: str | None = None
-) -> str:
-    query_params = {"referrer": "alert_email"}
-    if notification_uuid:
-        query_params.update({"notification_uuid": notification_uuid})
-
-    return organization.absolute_url(
-        f"/settings/{organization.slug}/integrations/{integration_slug}/",
-        query=urlencode(query_params),
-    )
-
-
-def get_issue_replay_link(group: Group, sentry_query_params: str = ""):
-    return str(group.get_absolute_url() + "replays/" + sentry_query_params)
-
-
-@dataclass
-class NotificationRuleDetails:
-    id: int
-    label: str
-    url: str
-    status_url: str
 
 
 def get_rules(
@@ -336,7 +251,7 @@ def occurrence_perf_to_email_html(context: Any) -> str:
 
 
 def get_spans(
-    entries: list[dict[str, list[dict[str, str | float]] | str]]
+    entries: list[dict[str, list[dict[str, str | float]] | str]],
 ) -> list[dict[str, str | float]] | None:
     """Get the given event's spans"""
     if not len(entries):
