+++ tests/sentry/notifications/test_helpers.py
@@ -12,11 +12,8 @@ from sentry.notifications.helpers import (
 )
 from sentry.notifications.models.notificationsettingoption import NotificationSettingOption
 from sentry.notifications.types import NotificationSettingEnum, NotificationSettingsOptionEnum
-from sentry.notifications.utils import (
-    get_email_link_extra_params,
-    get_group_settings_link,
-    get_rules,
-)
+from sentry.notifications.utils import get_rules
+from sentry.notifications.utils.links import get_email_link_extra_params, get_group_settings_link
 from sentry.silo.base import SiloMode
 from sentry.testutils.cases import TestCase
 from sentry.testutils.silo import assume_test_silo_mode, assume_test_silo_mode_of
